Wow! Minor bug in protocol buffers regarding the settedness of boolean fields (when set to false).  This causes the hk2 byte stream to be slightly different from the protocol buffer byte stream, but the data values are still the same.  Only the settedness of false valued boolean fields might be wrong

+++ hk2-configuration/persistence/hk2-xml/test1/src/main/java/org/glassfish/hk2/xml/test/pbuf/pc2/PBufPrecompileChild2.java
@@ -53,12 +53,15 @@ import org.jvnet.hk2.annotations.Contract;
  */
 @Contract
 @Hk2XmlPreGenerate
-@XmlType(propOrder={ "ZType", "BType", "types" })
+@XmlType(propOrder={
+        "ZType"
+        , "BType"
+        , "types" })
 @GenerateProto
 public interface PBufPrecompileChild2 {
     @XmlElement(name="btype")
-    public void setBType(byte jtype);
-    public byte getBType();
+    public void setBType(int jtype);
+    public int getBType();
     
     @XmlAttribute(required=true)
     public void setZType(boolean ztype);
